Split files; add cue color feedback, action toggle
