+++ package.json
@@ -3,7 +3,7 @@
 	"legacy": [
 		"qlab-advance"
 	],
-	"version": "1.1.8",
+	"version": "1.2.0",
 	"api_version": "1.0.0",
 	"keywords": [
 		"Software",
